configure.ac: more autoconf 2.70 compatibility work
Drop a bunch of obsolete or simply unnecessary checks so that we don't require
config.{guess,sub} (with autoconf <= 2.69 we also didn't need them, but that
likely meant cross compilation was silently broken or at least buggy).

Also reformat the list of headers and functions to check for, by putting each
on its own line and sorting alphabetically: that makes diffs much nicer when
we add/remove anything to those lists.